requirements.txt
starlette
httpx
authlib
itsdangerous
django
django-cors-headers
djangorestframework-api-key
django-filter
XlsxWriter
pandas
openpyxl
djangorestframework
django-sendgrid-v5
redis
celery
python-dotenv
sentry_sdk
channels
asyncio
channels-redis
uvicorn[standard]
nest_asyncio
Office365-REST-Python-Client
streamlit
streamlit-keycloak-lex
sqlalchemy
django_redis
django-storages[google]
DjangoSharepointStorage==1.1.3
click
django-lifecycle
python-keycloak==3.9.1
markdown
mistune
django-simple-history
markdown2
xhtml2pdf
tabulate
django-db-views
pre-commit
factory-boy
coverage
django-oauth2-codeflow
psycopg2-binary